+++ src/Profile/Message/HrvMessage.php
@@ -33,7 +33,7 @@ final class HrvMessage extends Message
     /**
      * Gets the time
      */
-    public function getTime(): int|null
+    public function getTime(): float|null
     {
         return $this->getFieldValue(0);
     }
